refactor(difficulty): put difficulty manager into its own class
